# manualDevelopment/backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: School Equipment Lending Portal API
  description: |

    ## Authentication
    All endpoints (except login and setup-password) require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## How to Authenticate in Swagger UI
    1. Click the **Authorize** button (top right)
    2. Login using POST /auth/login to get a token
    3. Copy the token from response
    4. In the Authorize dialog, paste the token (without "Bearer " prefix)
    5. Click Authorize

    ## User Roles
    - **Admin**: Full access to all resources
    - **Teacher**: Can approve/deny requests, view statistics
    - **Student**: Can request equipment, view own requests

    ## Base URL
    Replace `{BASE_URL}` with your API endpoint (e.g., `http://localhost:4000`)
  version: 1.0.0
  contact:
    name: API Support
    email: support@schoolequipment.com
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api
    description: Local server
  - url: https://school-equipment-landing-portal-q6q.vercel.app/api
    description: Production server (Vercel Live)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    EquipmentRequest:
      type: object
      properties:
        id:
          type: integer
          description: Request identifier
          example: 1
        user_id:
          type: integer
          description: ID of requesting user
          example: 5
        equipment_id:
          type: integer
          description: ID of requested equipment
          example: 3
        quantity:
          type: integer
          description: Quantity requested
          example: 2
        request_date:
          type: string
          format: date-time
          description: When request was made
          example: 2025-01-15T09:00:00Z
        required_date:
          type: string
          format: date
          description: When equipment is needed
          example: 2025-01-20
        return_date:
          type: string
          format: date
          description: When equipment should be returned
          example: 2025-01-25
        purpose:
          type: string
          description: Purpose of the request
          example: Class demonstration
        status:
          type: string
          enum: [Pending, Approved, Denied, Returned]
          description: Request status
          example: Pending
        denial_reason:
          type: string
          description: Reason for denial (if denied)
          example: Insufficient quantity available
        notes:
          type: string
          description: Admin notes
          example: Approved for demonstration class
        approved_date:
          type: string
          format: date-time
          description: When request was approved/denied
          example: 2025-01-15T10:30:00Z
        approved_by:
          type: integer
          description: ID of approver
          example: 2
        approved_by_name:
          type: string
          description: Name of approver
          example: Admin User
        requester_name:
          type: string
          description: Name of requester
          example: John Student
        requester_email:
          type: string
          description: Email of requester
          example: john@school.com
        equipment_name:
          type: string
          description: Name of requested equipment
          example: Microscope
        equipment_category:
          type: string
          description: Category of equipment
          example: Science Lab
    CreateEquipmentRequestRequest:
      type: object
      required:
        - equipment_id
        - quantity
        - required_date
        - purpose
      properties:
        equipment_id:
          type: integer
          description: ID of equipment to request
          example: 3
        quantity:
          type: integer
          minimum: 1
          description: Quantity needed
          example: 2
        required_date:
          type: string
          format: date
          description: When equipment is needed (YYYY-MM-DD)
          example: 2025-01-20
        return_date:
          type: string
          format: date
          description: When equipment will be returned (optional)
          example: 2025-01-25
        purpose:
          type: string
          description: Purpose for the equipment
          minLength: 5
          maxLength: 500
          example: "Biology class demonstration for topic: Plant Structure"
    ApproveRequestRequest:
      type: object
      properties:
        notes:
          type: string
          description: Optional notes from approver
          maxLength: 500
          example: Approved. Please ensure timely return.
    DenyRequestRequest:
      type: object
      required:
        - denial_reason
      properties:
        denial_reason:
          type: string
          description: Reason for denying the request
          minLength: 5
          maxLength: 500
          example: Equipment currently in use by another class
    RequestStats:
      type: object
      properties:
        total_requests:
          type: integer
          example: 45
        pending:
          type: integer
          example: 5
        approved:
          type: integer
          example: 30
        denied:
          type: integer
          example: 8
        returned:
          type: integer
          example: 2

    # Auth Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message describing what went wrong
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              value:
                type: string
paths:
...
  /requests:
    post:
      tags:
        - Equipment Requests
      summary: Create Equipment Request
      description: Submit a new equipment request. Students and Teachers only.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentRequestRequest"
            examples:
              basic:
                value:
                  equipment_id: 3
                  quantity: 2
                  required_date: "2025-01-20"
                  purpose: Biology class demonstration for cell structure topic
              withReturnDate:
                value:
                  equipment_id: 5
                  quantity: 1
                  required_date: "2025-01-22"
                  return_date: "2025-01-25"
                  purpose: Science fair project preparation
      responses:
        "201":
          description: Equipment request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Equipment request submitted successfully
        "400":
          description: Validation error or insufficient quantity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                insufficientQuantity:
                  value:
                    success: false
                    message: Insufficient quantity. Only 3 available
                duplicatePending:
                  value:
                    success: false
                    message: You already have a pending request for this equipment
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Equipment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/my-requests:
    get:
      tags:
        - Equipment Requests
      summary: Get My Requests
      description: Get your own equipment requests with optional status filter. Students and Teachers.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by request status
          schema:
            type: string
            enum: [Pending, Approved, Denied, Returned]
          example: Pending
      responses:
        "200":
          description: Requests retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EquipmentRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/all:
    get:
      tags:
        - Equipment Requests
      summary: Get All Requests
      description: Get all equipment requests with optional filters. Teachers and Admins only.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [Pending, Approved, Denied, Returned]
          example: Pending
        - name: user_id
          in: query
          description: Filter by requesting user
          schema:
            type: integer
          example: 5
        - name: equipment_id
          in: query
          description: Filter by equipment
          schema:
            type: integer
          example: 3
      responses:
        "200":
          description: Requests retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EquipmentRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Teacher or Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/{id}/approve:
    put:
      tags:
        - Equipment Requests
      summary: Approve Request
      description: Approve an equipment request. Teachers and Admins only.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Request ID to approve
          schema:
            type: integer
          example: 5
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRequestRequest"
      responses:
        "200":
          description: Request approved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Request approved successfully
        "400":
          description: Request already processed or insufficient quantity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                alreadyApproved:
                  value:
                    success: false
                    message: Request is already Approved
                insufficientQuantity:
                  value:
                    success: false
                    message: Insufficient equipment quantity available
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Teacher or Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/{id}/deny:
    put:
      tags:
        - Equipment Requests
      summary: Deny Request
      description: Deny an equipment request with a reason. Teachers and Admins only.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Request ID to deny
          schema:
            type: integer
          example: 5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DenyRequestRequest"
      responses:
        "200":
          description: Request denied successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Request denied successfully
        "400":
          description: Request already processed or missing denial reason
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Teacher or Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/{id}:
    delete:
      tags:
        - Equipment Requests
      summary: Cancel Request
      description: Cancel your own pending request. Can only cancel requests you created.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Request ID to cancel
          schema:
            type: integer
          example: 5
      responses:
        "200":
          description: Request cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Request cancelled successfully
        "400":
          description: Only pending requests can be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Can only cancel your own requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/statistics:
    get:
      tags:
        - Equipment Requests
      summary: Get Request Statistics
      description: Get statistics about equipment requests. Teachers and Admins only.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/RequestStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Teacher or Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"